Revert "Tiene que ver con el entorno de Pago, solo falta la validación (#20)"
This reverts commit d62230d6f3bbb5327034e4c8b64cf90a613cb6a7.
