+++ app/Http/Controllers/PagoController.php
@@ -4,15 +4,15 @@ namespace App\Http\Controllers;
 
 use App\Models\Alumno;
 use App\Models\ConceptoPago;
+use App\Models\DetallePago;
 use App\Models\Deuda;
 use App\Models\Pago;
-use App\Models\DetallePago;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
-use Carbon\Carbon;
 
 class PagoController extends Controller
 {
+
+    
     private static function doSearch($sqlColumns, $search, $maxEntriesShow){
         if (!isset($search)){
             $query = Pago::where('estado', '=', '1')->paginate($maxEntriesShow);
@@ -102,6 +102,8 @@ class PagoController extends Controller
 
     }
 
+
+
     public function create(Request $request){
         $alumnos = Alumno::all(['id_alumno', 'codigo_educando']);
         $deudas = Deuda::all(['id_deuda', 'id_alumno', 'id_concepto', 'periodo', 'monto_total']);
@@ -116,89 +118,54 @@ class PagoController extends Controller
         return view('gestiones.pago.create', compact('data'));
     }
 
-    public function createNewEntry(Request $request)
-    {
+    public function createNewEntry(Request $request){
         $request->validate([
-            'codigo_alumno'   => 'required|string',
             'id_deuda' => 'required|exists:deudas,id_deuda',
-            'metodo_pago' => 'required|array',
-            'metodo_pago.*' => 'required|string',
-            'detalle_recibo.*' => 'required|string',
-            'detalle_monto.*' => 'required|numeric|min:0.01',
-            'detalle_fecha.*' => 'required|date',
-            'detalle_observaciones.*' => 'nullable|string',
-            'voucher_path.*' => 'nullable|file|mimes:jpg,jpeg,png,pdf|max:2048',
-        ], [
-            'codigo_alumno.required' => 'Ingrese el código del educando.',
+            'detalle_fecha' => 'required|array',
+            'detalle_monto' => 'required|array',
+            'detalle_recibo' => 'required|array',
+            'detalle_observaciones' => 'required|array',
+        ],[
             'id_deuda.required' => 'Seleccione una deuda.',
-            'id_deuda.exists' => 'La deuda seleccionada no es válida.',
-            'metodo_pago.required' => 'Debe seleccionar un método de pago.',
-            'metodo_pago.*.required' => 'Seleccione un método de pago para cada detalle.',
-            'detalle_recibo.*.required' => 'Ingrese el número de operación.',
-            'detalle_monto.*.required' => 'Ingrese el monto pagado.',
-            'detalle_monto.*.numeric' => 'El monto debe ser numérico.',
-            'detalle_monto.*.min' => 'El monto debe ser mayor a 0.',
-            'detalle_fecha.*.required' => 'Ingrese la fecha del pago.',
-            'detalle_fecha.*.date' => 'La fecha ingresada no es válida.',
-            'detalle_observaciones.*.string' => 'La observación debe ser texto.',
-            'voucher_path.*.file' => 'Debe subir un archivo válido.',
-            'voucher_path.*.mimes' => 'Solo se permiten archivos JPG, JPEG, PNG o PDF.',
-            'voucher_path.*.max' => 'El archivo no puede superar 2MB.',
+            'detalle_fecha.required' => 'Agregue al menos un detalle de pago.',
         ]);
 
-        if (count($request->input('detalle_monto', [])) > 2) {
-            return back()->withErrors([
-                'detalle_monto' => 'Solo puede registrar hasta 2 detalles de pago.'
-            ])->withInput();
+        $idDeuda = $request->input('id_deuda');
+        $fechas = $request->input('detalle_fecha');
+        $montos = $request->input('detalle_monto');
+        $recibos = $request->input('detalle_recibo');
+        $observaciones = $request->input('detalle_observaciones');
+
+        // 1. Sumar todos los montos
+        $montoTotalPago = collect($montos)->map(function($m){ return floatval($m); })->sum();
+
+        // 2. Crear el pago principal
+        $pago = Pago::create([
+            'id_deuda' => $idDeuda,
+            'fecha_pago' => end($fechas), // última fecha como fecha del pago principal
+            'monto' => $montoTotalPago,
+            'observaciones' => $request->input('observaciones'),
+            // agrega otros campos si es necesario
+        ]);
+
+        // 3. Crear los detalles de pago
+        foreach ($fechas as $i => $fecha) {
+            DetallePago::create([
+                'id_pago' => $pago->getKey(),
+                'fecha_pago' => $fecha,
+                'monto' => floatval($montos[$i]),
+                'nro_recibo' => $recibos[$i],
+                'observacion' => $observaciones[$i],
+            ]);
         }
 
-        DB::transaction(function() use ($request) {
-            $idDeuda = $request->input('id_deuda');
-            $fechas = $request->input('detalle_fecha');
-            $montos = $request->input('detalle_monto');
-            $recibos = $request->input('detalle_recibo');
-            $observaciones = $request->input('detalle_observaciones');
-            $metodos = $request->input('metodo_pago');
+        // 4. Actualizar la deuda
+        $deuda = \App\Models\Deuda::find($idDeuda);
+        $deuda->monto_adelantado = floatval($deuda->monto_adelantado) + $montoTotalPago;
+        $deuda->monto_a_cuenta = floatval($deuda->monto_total) - floatval($deuda->monto_adelantado);
+        $deuda->save();
 
-            $montoTotalPago = collect($montos)->map(fn($m) => floatval($m))->sum();
-
-            $pago = Pago::create([
-                'id_deuda' => $idDeuda,
-                'fecha_pago' => end($fechas),
-                'monto' => $montoTotalPago,
-                'observaciones' => $request->input('observaciones'),
-                'estado_validacion' => 'pendiente',
-            ]);
-
-            foreach ($fechas as $i => $fecha) {
-                $metodo = strtolower(trim($metodos[$i] ?? ''));
-                $recibo = $recibos[$i] ?? null;
-                $monto = floatval($montos[$i] ?? 0);
-                $obs = $observaciones[$i] ?? null;
-
-                $voucherPath = null;
-                if (in_array($metodo, ['transferencia', 'yape'])) {
-                    if (!$request->hasFile("voucher_path.$i")) {
-                        throw new \Exception("Debe subir la constancia de pago para el método: $metodo");
-                    }
-                    $voucherPath = $request->file("voucher_path.$i")->store('vouchers', 'public');
-                }
-
-                DetallePago::create([
-                    'id_pago' => $pago->getKey(),
-                    'fecha_pago' => $fecha,
-                    'monto' => $monto,
-                    'nro_recibo' => $recibo,
-                    'observacion' => $obs,
-                    'metodo_pago' => $metodo,
-                    'voucher_path' => $voucherPath,
-                    'estado_validacion' => 'pendiente',
-                ]);
-            }
-        });
-
-        return redirect()->route('pago_view', ['created' => true])
-                        ->with('success', 'Pago registrado correctamente en estado pendiente.');
+        return redirect(route('pago_view', ['created' => true]));
     }
 
     public function edit(Request $request, $id)
@@ -264,38 +231,17 @@ class PagoController extends Controller
 
     public function viewDetalles($id)
     {
-        if (!isset($id)) {
+        if (!isset($id)){
             return redirect(route('pago_view'));
         }
 
         $pago = Pago::findOrFail($id);
+        $detalles = \App\Models\DetallePago::where('id_pago', $id)->get();
 
-        $detalles = DetallePago::where('id_pago', $id)->get();
-
-        $alumno = Alumno::select(
-            'codigo_educando',
-            'codigo_modular',
-            'año_ingreso',
-            'dni',
-            'apellido_paterno',
-            'apellido_materno',
-            'primer_nombre',
-            'otros_nombres',
-            'sexo',
-            'fecha_nacimiento',
-            'direccion',
-            'telefono'
-        )
-        ->where('id_alumno', function($q) use ($pago) {
-            $q->select('id_alumno')
-            ->from('deudas')
-            ->where('id_deuda', $pago->id_deuda)
-            ->limit(1);
-        })
-        ->first();
-
-        return view('gestiones.pago.detalles', compact('pago', 'detalles', 'alumno'));
+        return view('gestiones.pago.detalles', compact('pago', 'detalles'));
     }
+
+
 
     public function buscarAlumno($codigo)
     {
@@ -310,38 +256,13 @@ class PagoController extends Controller
 
         $deudas = Deuda::where('id_alumno', $alumno->id_alumno)->get();
 
+        // Obtener deudas con contenido específico
         $deudasFormateadas = $deudas->map(function ($deuda) {
-            $pagos = Pago::where('id_deuda', $deuda->id_deuda)->get();
-
-            $pagoIds = $pagos->pluck('id_pago')->toArray();
-            $detalles = collect();
-            if (!empty($pagoIds)) {
-                $detalles = DetallePago::whereIn('id_pago', $pagoIds)->get();
-            }
-
-            $montoPagado = $detalles->sum(function ($d) {
-                return floatval($d->monto);
-            });
-
-            $detallesArr = $detalles->map(function ($d) {
-                return [
-                    'metodo_pago'      => $d->metodo_pago,
-                    'nro_recibo'       => $d->nro_recibo,
-                    'monto'            => floatval($d->monto),
-                    'fecha_pago'       => $d->fecha_pago ? $d->fecha_pago->format('Y-m-d H:i:s') : null,
-                    'observacion'      => $d->observacion,
-                    'voucher_path'     => $d->voucher_path,
-                    'estado_validacion'=> $d->estado_validacion,
-                ];
-            })->values();
-
             return [
-                'id_deuda'      => $deuda->id_deuda,
-                'periodo'       => $deuda->periodo,
-                'monto_total'   => floatval($deuda->monto_total),
-                'monto_pagado'  => round($montoPagado, 2),
-                'concepto'      => $deuda->conceptoPago ? $deuda->conceptoPago->descripcion : 'Sin concepto',
-                'detalles_pago' => $detallesArr,
+                'id_deuda' => $deuda->id_deuda,
+                'periodo' => $deuda->periodo,
+                'monto_total' => $deuda->monto_total,
+                'concepto' => $deuda->conceptoPago ? $deuda->conceptoPago->descripcion : 'Sin concepto'
             ];
         });
 
@@ -350,9 +271,11 @@ class PagoController extends Controller
             'alumno' => [
                 'id_alumno' => $alumno->id_alumno,
                 'codigo_educando' => $alumno->codigo_educando,
-                'nombre_completo' => trim($alumno->apellido_paterno . ' ' . $alumno->apellido_materno . ' ' . $alumno->primer_nombre)
+                'nombre_completo' => $alumno->apellido_paterno . ' ' . $alumno->apellido_materno . ' ' . $alumno->primer_nombre
             ],
-            'deudas' => $deudasFormateadas,
+            'deudas' => $deudasFormateadas
         ]);
     }
+
+
 }
