+++ app/Providers/AppServiceProvider.php
@@ -38,19 +38,22 @@ use App\Observers\UserObserver;
 use Illuminate\Pagination\Paginator;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\ServiceProvider;
-use Illuminate\Support\Facades\Schema;
 
 class AppServiceProvider extends ServiceProvider
 {
+    /**
+     * Register any application services.
+     */
     public function register(): void
     {
         //
     }
 
+    /**
+     * Bootstrap any application services.
+     */
     public function boot(): void
     {
-        Schema::defaultStringLength(191);
-
         $this->registerObservers();
 
         Paginator::useBootstrapFour();
@@ -69,6 +72,7 @@ class AppServiceProvider extends ServiceProvider
         Gate::define('manage-resource', function (User $user, $resource, $action = 'view'){
             return $this->hasResourcePermissions($user, $resource, $action);
         });
+
     }
 
     private function registerObservers(){
