+++ database/seeders/ConceptosSeeder.php
@@ -20,11 +20,14 @@ use Illuminate\Database\Seeder;
 use Faker\Factory as FakerFactory;
 class ConceptosSeeder extends Seeder
 {
-    
+    /**
+     * Seed the application's database.
+     */
     public function run(): void
     {
         // Deshabilitamos temporalmente el registro de acciones, ya que estamos ejecutando un seeder.
         LogsActions::disable();
+
 
         ConceptoAccion::create(
             [
@@ -89,6 +92,11 @@ class ConceptosSeeder extends Seeder
             $datosdepartamentos[] = $nuevodepartamento;
         }
 
+
+
+
+
+
         $niveleseducativos = [
             [
                 'nombre_nivel'=>'Inicial',
@@ -115,6 +123,7 @@ class ConceptosSeeder extends Seeder
             );
             $datosniveleseducativos[] = $nuevoniveleducativo;
         }
+
 
         $datosgrados = [];
 
@@ -219,6 +228,7 @@ class ConceptosSeeder extends Seeder
             }
             
         }
+
 
         $añosescolares = [
             2025,2026
@@ -357,7 +367,14 @@ class ConceptosSeeder extends Seeder
             }
         }
 
+
+
+
+
         // Restablecemos el registro de acciones.
         LogsActions::enable();
     }
+
+
+
 }
