+++ resources/views/components/forms/btn-generar.blade.php
@@ -1,16 +1,11 @@
-<!-- Botón Buscar Deuda -->
-<button type="button" id="btn-generar"
-    class="flex items-center rounded-lg border border-amber-300 bg-amber-200 overflow-hidden focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-amber-400">
-
-    <span class="w-full px-4 py-2 text-sm font-semibold text-gray-800">
-        Buscar Deuda
-    </span>
-
-    <span class="px-3 py-2 bg-amber-300 border-l border-amber-400 flex items-center justify-center">
-        <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24"
-            stroke="currentColor" class="h-5 w-5 text-gray-700">
-            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
-                d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z" />
-        </svg>
-    </span>
+<button
+  type="button"
+  id="btn-generar"
+  class="flex flex-row justify-center gap-2 rounded-lg px-4 py-2 text-sm font-semibold focus:outline-none focus:ring-2 focus:ring-offset-2 transition bg-amber-200"
+>
+  <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor" class="h-4 w-4">
+    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
+      d="M3 18v-6a3 3 0 013-3h2V7a3 3 0 116 0v2h2a3 3 0 013 3v6m-6 0V9m0 9v3m-6-3v3" />
+  </svg>
+  Generar Deudas
 </button>
