--- resources/views/gestiones/validacion_pago/edit.blade.php
@@ -1,54 +1,0 @@
-@extends('base.administrativo.blank')
-
-@section('titulo')
-  Pagos | Edición
-@endsection
-
-@section('contenido')
-  <div class="p-8 m-4 bg-gray-100 dark:bg-white/[0.03] rounded-2xl">
-    <div class="flex pb-4 justify-between items-center">
-      <h2 class="text-lg dark:text-gray-200 text-gray-800">Estás editando el Pago con ID {{ $data['id'] }}</h2>
-
-      <div class="flex gap-4">
-        <input form="form" target="" type="submit" form=""
-          class="cursor-pointer inline-flex items-center gap-2 rounded-lg border border-gray-300 bg-gray-200 px-4 py-2.5 text-theme-sm font-medium text-gray-700 shadow-theme-xs hover:bg-gray-300 hover:text-gray-800 dark:border-gray-700 dark:bg-gray-600 dark:text-gray-200 dark:hover:bg-white/[0.03] dark:hover:text-gray-200"
-          value="Guardar"
-        >
-
-        <a
-          href="{{ $data["return"] }}"
-          class="inline-flex items-center gap-2 rounded-lg border border-gray-300 bg-gray-200 px-4 py-2.5 text-theme-sm font-medium text-gray-700 shadow-theme-xs hover:bg-gray-300 hover:text-gray-800 dark:border-gray-700 dark:bg-gray-600 dark:text-gray-200 dark:hover:bg-white/[0.03] dark:hover:text-gray-200"
-        >
-          Cancelar
-        </a>
-      </div>
-    </div>
-
-    <form method="POST" id="form" class="flex flex-col gap-4" action="">
-      @method('PATCH')
-      @csrf
-
-      @include('components.forms.datetime-picker',[
-        'label' => 'Fecha pago',
-        'error' => $errors->first(Str::snake('Fecha pago')) ?? false,
-        'value' => old(Str::snake('Fecha pago')) ?? $data['default']['fecha_pago']
-      ])
-
-      @include('components.forms.string',[
-        'label' => 'Monto',
-        'error' => $errors->first(Str::snake('Monto')) ?? false,
-        'value' => old(Str::snake('Monto')) ?? $data['default']['monto']
-      ])
-
-      @include('components.forms.text-area',[
-        'label' => 'Observaciones',
-        'error' => $errors->first(Str::snake('Observaciones')) ?? false,
-        'value' => old(Str::snake('Observaciones')) ?? $data['default']['observaciones']
-      ])
-
-    </form>
-@endsection
-
-@section('custom-js')
-  <script src="{{ asset('js/tables.js') }}"></script>
-@endsection
