+++ resources/views/layout/sidebar.blade.php
@@ -39,7 +39,7 @@
             class="menu-group-title"
             :class="sidebarToggle ? 'lg:hidden' : ''"
           >
-            MENU 
+            MENU
           </span>
 
           <svg
@@ -62,7 +62,7 @@
 
         <ul class="flex flex-col gap-4 mb-6">
 
-        @yield(section: 'opciones')
+        @yield('opciones')
         </ul>
       </div>
     </nav>
