--- routes/administrativa/validacion_pago.php
@@ -1,6 +1,0 @@
-<?php
-
-use App\Http\Controllers\ValidacionPagoController;
-
-Route::get('/', [ValidacionPagoController::class, 'index'])
-        ->name('view');
